E2E test restructuring (#377)
* Renaming and reorganizing E2E test

- unifying naming structure
- creating folders

* E2E test refactors

First round of refactorings

RoleAuthorizerE2E done. Should serve as example for the rest

finsha Authorizers refactor and small subdivision changes

add setUp func + moduleConfigurations to all tests

add template to E2ERegistry

Finish rest of easy refactors

Also deleted RegisterModuleAtFactory.t. sol because it was duplicating the setUp tests in the E2EModuleRegsitry

Orchestrator E2E refactor

* clean up imports + small final refactor

* add template + pre-commit

* Uncomment template so linter doesn't complain

* remove TODOs

* change file format for template

tests were failing

* new fmt in foundry

---------

+++ test/modules/paymentProcessor/StreamingPaymentProcessor.t.sol
@@ -1044,7 +1044,7 @@ contract StreamingPaymentProcessorTest is ModuleTest {
         paymentProcessor.cancelRunningPayments(paymentClient);
 
         // measure recipients balances before attempting second claim.
-        uint[] memory balancesBefore = new uint256[](recipients.length);
+        uint[] memory balancesBefore = new uint[](recipients.length);
         for (uint i; i < recipients.length; i++) {
             vm.prank(recipients[i]);
             balancesBefore[i] = _token.balanceOf(recipients[i]);
